+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 ### 🚀 Features
 
 - Added possibility to use subfields in show utils and added padding option to the YoloOnnxDetector
+- Added SignatureDetector
+- Updated ImageCropBoxes to support multiple boxes
 - Added LineOrientation detector model to the TesseractRecognizer
 - Added FaceDetector, SignatureDetector
 - Added PdfAssembler
+
+### 🐛 Bug Fixes
+
+- Fixed borders in show utils